requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
motor>=3.4,<4
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
python-jose[cryptography]>=3.3,<4
bcrypt>=4.0,<5
aiosmtplib>=3.0,<4
python-dotenv>=1.0,<2
email-validator>=2.1,<3
boto3>=1.34,<2
python-multipart>=0.0.9,<1
